Add THE-VALUES HIR instruction and fix complex values types
Previously, THE-ASTs compiled in a context expecting a values location
attempted to expand the location with M->F and insert THE instructions,
which operate on lexical locations. I found that this inhibited type
propagation because said lexical locations would often be used only for
the THE instructions - e.g., if a function ended with
(the whatever (funcall ...))
the values from the funcall would be split up by THE, but those lexical
locations would be useful as the return instruction would just use the
values location.

Also, the splitting isn't possible in general, due to &OPTIONAL and
&REST in values types, i.e. unknown-values returns.